# bearer-token-service.v2/docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Bearer Token Service V2
  description: |
    认证服务 API

    ## 认证方式

    ### HMAC 签名认证
    所有账户管理和 Token 管理 API 都需要使用 HMAC-SHA256 签名认证。

    #### 签名步骤
    1. 构建待签名字符串: `<HTTP_METHOD>\n<URI_PATH>\n<TIMESTAMP>\n<REQUEST_BODY>`
    2. 使用 SecretKey 生成 HMAC-SHA256 签名
    3. 构造 Authorization Header: `QINIU <ACCESS_KEY>:<SIGNATURE>`

    ### Qstub 认证
    七牛内部服务使用 Qstub Token 进行认证。
  version: 2.0.0
  contact:
    name: API Support
    email: support@example.com
  license:
    name: MIT
servers:
  - url: http://localhost:8080
    description: 本地开发环境
  - url: https://api.example.com
    description: 生产环境

tags:
  - name: 账户管理
    description: 账户注册、信息查询、密钥管理
  - name: Token 管理
    description: Bearer Token 的创建、查询、更新、删除
  - name: Token 验证
    description: Bearer Token 的验证和权限检查
  - name: 审计日志
    description: 操作审计日志查询
  - name: 健康检查
    description: 服务健康状态检查

paths:
  /health:
    get:
      summary: 健康检查
      description: 检查服务运行状态
      tags:
        - 健康检查
      responses:
        '200':
          description: 服务正常运行
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok

  /api/v2/accounts/register:
    post:
      summary: 注册账户
      description: 创建新的租户账户并获取 AccessKey/SecretKey
      tags:
        - 账户管理
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - company
                - password
              properties:
                email:
                  type: string
                  format: email
                  description: 邮箱地址
                  example: user@example.com
                company:
                  type: string
                  description: 公司名称
                  example: Example Inc
                password:
                  type: string
                  format: password
                  description: 密码
                  example: securePassword123
      responses:
        '201':
          description: 账户注册成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  account_id:
                    type: string
                    example: acc_1a2b3c4d5e6f
                  email:
                    type: string
                    example: user@example.com
                  company:
                    type: string
                    example: Example Inc
                  access_key:
                    type: string
                    example: AK_f8e7d6c5b4a392817a6b5c4d3e2f1a0b9c8d7e6f5a4b3c2d1e0f9a8b7c6d5e4
                  secret_key:
                    type: string
                    description: 仅在注册时返回一次，请妥善保存
                    example: SK_a1b2c3d4e5f6g7h8i9j0k1l2m3n4o5p6q7r8s9t0u1v2w3x4y5z6a7b8c9d0e1f2
                  created_at:
                    type: string
                    format: date-time
                    example: 2025-12-25T10:00:00Z
        '400':
          $ref: '#/components/responses/BadRequest'

  /api/v2/accounts/me:
    get:
      summary: 获取账户信息
      description: 获取当前登录账户的信息
      tags:
        - 账户管理
      security:
        - HMACAuth: []
        - QstubAuth: []
      responses:
        '200':
          description: 账户信息获取成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountInfo'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /api/v2/accounts/regenerate-sk:
    post:
      summary: 重新生成 SecretKey
      description: 出于安全考虑，定期轮换 SecretKey。旧的 SecretKey 立即失效。
      tags:
        - 账户管理
      security:
        - HMACAuth: []
        - QstubAuth: []
      responses:
        '200':
          description: SecretKey 重新生成成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_key:
                    type: string
                    example: AK_f8e7d6c5b4a392817a6b5c4d3e2f1a0b9c8d7e6f5a4b3c2d1e0f9a8b7c6d5e4
                  secret_key:
                    type: string
                    description: 新的 SecretKey，仅显示一次
                    example: SK_b2c3d4e5f6g7h8i9j0k1l2m3n4o5p6q7r8s9t0u1v2w3x4y5z6a7b8c9d0e1f2a3
                  updated_at:
                    type: string
                    format: date-time
                    example: 2025-12-25T10:30:00Z
        '401':
          $ref: '#/components/responses/Unauthorized'

  /api/v2/tokens:
    post:
      summary: 创建 Bearer Token
      description: 创建一个新的 Bearer Token，并指定权限范围（Scope）
      tags:
        - Token 管理
      security:
        - HMACAuth: []
        - QstubAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - description
                - scope
              properties:
                description:
                  type: string
                  description: Token 描述
                  example: Production read-only token
                scope:
                  type: array
                  description: 权限范围，支持通配符
                  items:
                    type: string
                  example: ["storage:read", "cdn:refresh"]
                expires_in_seconds:
                  type: integer
                  description: 过期时间（秒），0 表示永不过期
                  example: 7776000
                prefix:
                  type: string
                  description: 自定义 Token 前缀（默认 sk-）
                  example: custom_bearer_
                rate_limit:
                  type: object
                  description: API 频率限制
                  properties:
                    requests_per_minute:
                      type: integer
                      example: 1000
      responses:
        '201':
          description: Token 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Token'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

    get:
      summary: 列出 Tokens
      description: 列出当前账户的所有 Tokens
      tags:
        - Token 管理
      security:
        - HMACAuth: []
        - QstubAuth: []
      parameters:
        - name: active_only
          in: query
          description: 仅显示激活的 Token
          schema:
            type: boolean
            default: false
        - name: limit
          in: query
          description: 返回数量（最大 100）
          schema:
            type: integer
            default: 50
            maximum: 100
        - name: offset
          in: query
          description: 偏移量
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Token 列表获取成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  account_id:
                    type: string
                    example: acc_1a2b3c4d5e6f
                  tokens:
                    type: array
                    items:
                      $ref: '#/components/schemas/TokenSummary'
                  total:
                    type: integer
                    example: 1
        '401':
          $ref: '#/components/responses/Unauthorized'

  /api/v2/tokens/{token_id}:
    get:
      summary: 获取 Token 详情
      description: 获取单个 Token 的详细信息
      tags:
        - Token 管理
      security:
        - HMACAuth: []
        - QstubAuth: []
      parameters:
        - name: token_id
          in: path
          required: true
          description: Token ID
          schema:
            type: string
            example: tk_9z8y7x6w5v4u
      responses:
        '200':
          description: Token 详情获取成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenSummary'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      summary: 删除 Token
      description: 永久删除 Token
      tags:
        - Token 管理
      security:
        - HMACAuth: []
        - QstubAuth: []
      parameters:
        - name: token_id
          in: path
          required: true
          description: Token ID
          schema:
            type: string
            example: tk_9z8y7x6w5v4u
      responses:
        '200':
          description: Token 删除成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Token deleted successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  /api/v2/tokens/{token_id}/status:
    put:
      summary: 更新 Token 状态
      description: 启用或禁用 Token
      tags:
        - Token 管理
      security:
        - HMACAuth: []
        - QstubAuth: []
      parameters:
        - name: token_id
          in: path
          required: true
          description: Token ID
          schema:
            type: string
            example: tk_9z8y7x6w5v4u
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - is_active
              properties:
                is_active:
                  type: boolean
                  description: Token 激活状态
                  example: false
      responses:
        '200':
          description: Token 状态更新成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  token_id:
                    type: string
                    example: tk_9z8y7x6w5v4u
                  is_active:
                    type: boolean
                    example: false
                  updated_at:
                    type: string
                    format: date-time
                    example: 2025-12-25T10:30:00Z
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  /api/v2/tokens/{token_id}/stats:
    get:
      summary: 获取 Token 使用统计
      description: 查看 Token 的使用情况
      tags:
        - Token 管理
      security:
        - HMACAuth: []
        - QstubAuth: []
      parameters:
        - name: token_id
          in: path
          required: true
          description: Token ID
          schema:
            type: string
            example: tk_9z8y7x6w5v4u
      responses:
        '200':
          description: Token 统计信息获取成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  token_id:
                    type: string
                    example: tk_9z8y7x6w5v4u
                  total_requests:
                    type: integer
                    example: 12567
                  last_used_at:
                    type: string
                    format: date-time
                    example: 2025-12-25T09:45:00Z
                  created_at:
                    type: string
                    format: date-time
                    example: 2025-12-25T10:00:00Z
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  /api/v2/validate:
    post:
      summary: 验证 Bearer Token
      description: 验证 Token 的有效性，并可选地检查特定权限
      tags:
        - Token 验证
      security:
        - BearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                required_scope:
                  type: string
                  description: 可选，要求的权限
                  example: storage:read
      responses:
        '200':
          description: Token 验证成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  valid:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Token is valid
                  token_info:
                    type: object
                    properties:
                      token_id:
                        type: string
                        example: tk_9z8y7x6w5v4u
                      account_id:
                        type: string
                        example: acc_1a2b3c4d5e6f
                      scope:
                        type: array
                        items:
                          type: string
                        example: ["storage:read", "cdn:refresh"]
                      is_active:
                        type: boolean
                        example: true
                      expires_at:
                        type: string
                        format: date-time
                        example: 2026-03-25T10:00:00Z
                  permission_check:
                    type: object
                    properties:
                      requested:
                        type: string
                        example: storage:read
                      granted:
                        type: boolean
                        example: true
        '401':
          description: Token 验证失败
          content:
            application/json:
              schema:
                type: object
                properties:
                  valid:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: Token has expired

  /api/v2/audit-logs:
    get:
      summary: 查询审计日志
      description: 查询当前账户的操作审计日志
      tags:
        - 审计日志
      security:
        - HMACAuth: []
        - QstubAuth: []
      parameters:
        - name: action
          in: query
          description: 过滤操作类型
          schema:
            type: string
            example: create_token
        - name: resource_id
          in: query
          description: 过滤资源 ID
          schema:
            type: string
        - name: start_time
          in: query
          description: 开始时间（ISO 8601）
          schema:
            type: string
            format: date-time
        - name: end_time
          in: query
          description: 结束时间（ISO 8601）
          schema:
            type: string
            format: date-time
        - name: limit
          in: query
          description: 返回数量
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          description: 偏移量
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: 审计日志获取成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  account_id:
                    type: string
                    example: acc_1a2b3c4d5e6f
                  logs:
                    type: array
                    items:
                      $ref: '#/components/schemas/AuditLog'
                  total:
                    type: integer
                    example: 1
        '401':
          $ref: '#/components/responses/Unauthorized'

components:
  securitySchemes:
    HMACAuth:
      type: apiKey
      in: header
      name: Authorization
      description: |
        HMAC 签名认证

        格式: `QINIU <ACCESS_KEY>:<SIGNATURE>`

        需要同时提供 X-Qiniu-Date 头部

    QstubAuth:
      type: http
      scheme: bearer
      description: |
        Qstub 认证（七牛内部服务）

        格式: `Bearer <QSTUB_TOKEN>`

    BearerAuth:
      type: http
      scheme: bearer
      description: |
        Bearer Token 认证

        格式: `Bearer <TOKEN>`

  schemas:
    AccountInfo:
      type: object
      properties:
        id:
          type: string
          example: acc_1a2b3c4d5e6f
        email:
          type: string
          example: user@example.com
        company:
          type: string
          example: Example Inc
        access_key:
          type: string
          example: AK_f8e7d6c5b4a392817a6b5c4d3e2f1a0b9c8d7e6f5a4b3c2d1e0f9a8b7c6d5e4
        status:
          type: string
          example: active
        created_at:
          type: string
          format: date-time
          example: 2025-12-25T10:00:00Z
        updated_at:
          type: string
          format: date-time
          example: 2025-12-25T10:00:00Z

    Token:
      type: object
      properties:
        token_id:
          type: string
          example: tk_9z8y7x6w5v4u
        token:
          type: string
          example: custom_bearer_a1b2c3d4e5f6g7h8i9j0k1l2m3n4o5p6q7r8s9t0u1v2w3x4y5z6a7b8c9d0e1f2
        account_id:
          type: string
          example: acc_1a2b3c4d5e6f
        description:
          type: string
          example: Production read-only token
        scope:
          type: array
          items:
            type: string
          example: ["storage:read", "cdn:refresh"]
        rate_limit:
          type: object
          properties:
            requests_per_minute:
              type: integer
              example: 1000
        created_at:
          type: string
          format: date-time
          example: 2025-12-25T10:00:00Z
        expires_at:
          type: string
          format: date-time
          example: 2026-03-25T10:00:00Z
        is_active:
          type: boolean
          example: true

    TokenSummary:
      type: object
      properties:
        token_id:
          type: string
          example: tk_9z8y7x6w5v4u
        token_preview:
          type: string
          example: sk-a1b2c3d4e5f6g7******************************c9d0e1f2
        description:
          type: string
          example: Production read-only token
        scope:
          type: array
          items:
            type: string
          example: ["storage:read", "cdn:refresh"]
        created_at:
          type: string
          format: date-time
          example: 2025-12-25T10:00:00Z
        expires_at:
          type: string
          format: date-time
          example: 2026-03-25T10:00:00Z
        is_active:
          type: boolean
          example: true
        total_requests:
          type: integer
          example: 12567
        last_used_at:
          type: string
          format: date-time
          example: 2025-12-25T09:45:00Z

    AuditLog:
      type: object
      properties:
        id:
          type: string
          example: log_xyz123
        account_id:
          type: string
          example: acc_1a2b3c4d5e6f
        action:
          type: string
          example: create_token
        resource_id:
          type: string
          example: tk_9z8y7x6w5v4u
        ip:
          type: string
          example: 203.0.113.42
        user_agent:
          type: string
          example: Mozilla/5.0...
        result:
          type: string
          example: success
        timestamp:
          type: string
          format: date-time
          example: 2025-12-25T10:00:00Z

    Error:
      type: object
      properties:
        code:
          type: integer
          example: 4001
        message:
          type: string
          example: Invalid signature
        details:
          type: string
          example: The provided signature does not match the expected signature
        request_id:
          type: string
          example: req_abc123

  responses:
    BadRequest:
      description: 请求参数错误
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: 400
            message: Invalid request parameters
            details: Missing required field 'description'
            request_id: req_abc123

    Unauthorized:
      description: 认证失败
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: 401
            message: Authentication failed
            details: Invalid or expired credentials
            request_id: req_abc123

    Forbidden:
      description: 权限不足
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: 403
            message: Permission denied
            details: Insufficient permissions for this operation
            request_id: req_abc123

    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: 404
            message: Resource not found
            details: The requested resource does not exist
            request_id: req_abc123

    TooManyRequests:
      description: 请求过于频繁
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: 429
            message: Too many requests
            details: Rate limit exceeded
            request_id: req_abc123